# packages/contract/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "REST API layanan inspeksi & sampling lapangan — Go + Gin dengan struktur modular ala NestJS.\nAutentikasi di luar cakupan PoC: seluruh endpoint terbuka dan peran dipilih di frontend.",
        "title": "Verifield API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "Mengambil daftar user dengan paginasi, pencarian, dan sorting",
                "tags": [
                    "users"
                ],
                "summary": "Daftar user",
                "parameters": [
                    {
                        "description": "Halaman",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "description": "Jumlah data per halaman",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "description": "Cari berdasarkan nama/email",
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Kolom sorting",
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "updated_at",
                                "name",
                                "email",
                                "role"
                            ]
                        }
                    },
                    {
                        "description": "Arah sorting",
                        "name": "sort_dir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.UserResponse"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Buat user baru",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.CreateUserDTO"
                            }
                        }
                    },
                    "description": "Data user",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.UserResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Detail user",
                "parameters": [
                    {
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.UserResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Hapus user",
                "parameters": [
                    {
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Perbarui user",
                "parameters": [
                    {
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.UpdateUserDTO"
                            }
                        }
                    },
                    "description": "Field yang ingin diubah",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_user_dto.UserResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "verifield-be_internal_common_apperror.FieldError": {
                "type": "object",
                "properties": {
                    "field": {
                        "type": "string",
                        "example": "email"
                    },
                    "message": {
                        "type": "string",
                        "example": "email wajib diisi"
                    }
                }
            },
            "verifield-be_internal_common_response.Envelope": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "data": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/verifield-be_internal_common_apperror.FieldError"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/verifield-be_internal_common_response.Meta"
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "verifield-be_internal_common_response.Meta": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "total_page": {
                        "type": "integer",
                        "example": 5
                    }
                }
            },
...
            "verifield-be_internal_modules_user_dto.CreateUserDTO": {
                "type": "object",
                "required": [
                    "email",
                    "name",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "maxLength": 160,
                        "example": "siti@verifield.id"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 3,
                        "example": "Siti Rahma"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8,
                        "example": "rahasia123"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "client",
                            "inspector",
                            "cs"
                        ],
                        "example": "client"
                    }
                }
            },
            "verifield-be_internal_modules_user_dto.UpdateUserDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "maxLength": 160,
                        "example": "siti@verifield.id"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 3,
                        "example": "Siti Rahma"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 8,
                        "example": "rahasia123"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "client",
                            "inspector",
                            "cs"
                        ],
                        "example": "admin"
                    }
                }
            },
            "verifield-be_internal_modules_user_dto.UserResponse": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "example": "siti@verifield.id"
                    },
                    "id": {
                        "type": "string",
                        "example": "6f1e6f0c-6f2a-4c5e-9f3a-0b6b1a4d2c11"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Siti Rahma"
                    },
                    "role": {
                        "type": "string",
                        "example": "client"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            }
        }
    }
}